impersonation: dashboards are specific to the user
This is a pretty big change, it gets rid of the
`User` resource and instead relies on k8s' native
authz concept. There are some threads that haven't
been fully taken care of yet - namely granting
permissions through the CLI doesn't work and
there's no documentation on what is required for
the source account.