feat(logs): default to showing previous logs
It is nice to be able to show logs from any
previous terminated containers, so default to
that. Unfortunately, if those logs don't exist
anymore, k8s fails to open the stream. This checks
the error message when creating a stream and stops
trying to get previous ones if they don't exist.