feat(widget): introduce `Container` trait
In preparation for doing animations, the adhoc
nature of "containers" and enabling/disabling
their components became somewhat non-tenable.
Combined with a lack of layers, it was time for a
new solution. The original `Widget` trait has been
kept the same - except for some additional new
methods that allow for dynamic layouts.

The `Container` trait provides a default
implementation of `Widget` that assumes there are
multiple widgets.

- The *first* widget is assumed to be terminal -
  if it gets exited, the entire container is
  exited.
- Widgets set their z-index (this should probably
  be changed). Each z-index is grouped together
  and the highest layer is rendered.
- When there are multiple widgets on a layer, the
  `placement` method is called. Implementing
  widgets are expected to say how they want to be
  placed vertically and horizontally.

Note: this is a first pass and has only been
implemented on the `Apex` widget. It'll need to
iterate as other widgets are converted.