feat(install): adding `resources install`
Kustomize was just too difficult to get working
well in getting started. This introduces an
`install` and `delete` command that does it for
all resources, not just the CRD. There's some
minimal patching that happens to the resources so
that they hopefully can still be applied with
`kubectl` directly from the repo.